Fix undefined method issue with Edit User form
the form was trying to call helpers which are specified by Devise and so aren't
present when you access a Devise form from a non-Devise controller.

To get around this, you need to add the following methods to the helper
class of the controller you wish to display the form under.
Alternatively, you can just add them to your application helper to make
them available anywhere.

http://stackoverflow.com/questions/4081744/devise-form-within-a-different-controller